fix: harden HTTP request parser with multiple security and robustness fixes
- Fix type mismatch: use ssize_t for bytes_readed to match read() return
- Add NULL safety check in httpparser_run for parser and buffer
- Reject spaces/tabs in HTTP method names per RFC 7230
- Handle Content-Length: 0 explicitly for proper pipelined request handling
- Add REDIRECT_BAD_REQUEST status for malformed URI redirects
- Fix httpparser_set_uri to handle OOM and bad request return codes
- Fix path traversal check to free memory before returning error
- Fix Connection header parsing to handle comma-separated tokens (RFC 7230)
- Fix memory leaks in Range and Cookie header parsing (free before reassign)
- Fix isdigit() call with unsigned char cast for safe char classification

+++ protocols/http/server/parsers/httprequestparser.h
@@ -43,7 +43,7 @@ void httpparser_init(httprequestparser_t* parser, connection_t* connection);
 void httpparser_free(void* arg);
 void httpparser_reset(httprequestparser_t*);
 int httpparser_run(httprequestparser_t* parser);
-void httpparser_set_bytes_readed(httprequestparser_t*, int);
+void httpparser_set_bytes_readed(httprequestparser_t*, size_t);
 void httpparser_prepare_continue(httprequestparser_t* parser);
 int httpparser_set_uri(httprequest_t*, const char*, size_t);
 void httpparser_append_query(httprequest_t*, query_t*);
